+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Intercom for Cordova/PhoneGap
 
-## 9.2.1 (2020-09-09)
+## 9.3.0 (2020-11-10)
+
+* The Intercom Cordova plugin has been updated to use v8.3.0 of the Intercom Android SDK and v8.1.0 of the iOS SDK as the latest version.
+* New feature: You can now open an article or Mobile Carousel programmatically in your app. We have added two new methods to our API to enable this, [displayArticle](https://developers.intercom.com/installing-intercom/docs/cordova-phonegap-configuration#section-present-an-article-programmatically) and [displayCarousel](https://developers.intercom.com/installing-intercom/docs/cordova-phonegap-configuration#section-present-a-carousel-programmatically).
+
+## 9.2.1 (2020-09-17)
 
 * The Intercom Cordova plugin has been updated to use v8.2.0 of the Intercom Android SDK.
 * The Intercom Cordova plugin has been updated to use v8.0.0 of the Intercom iOS SDK.